server/requirements.txt
chromadb
flask
flask-cors
python-dotenv
langchain-community
langchain-groq
langchain-cohere